Did some rearranging

+++ Core/dataHandling.c
@@ -1,12 +1,12 @@
-#include "dataHandling.h"
-#include "../Utils/logger.h"
+#include"Core/dataHandling.h"
+#include"Utils/logger.h"
 #include<WinSock2.h>
 #include<stdlib.h>
 
 int sendData(SOCKET *csock, char* buf)
 {
-    int bytes_sent = send(*csock, buf, strlen(buf)*sizeof(char), 0);
-    if(bytes_sent == SOCKET_ERROR)
+    size_t bytes_sent = send(*csock, buf, strlen(buf)*sizeof(char), 0);
+    if(bytes_sent == (size_t) SOCKET_ERROR) 
     {
         logError("Coudln't send Response.");
         return 0;
